requirements.txt
fastapi
uvicorn
faiss-cpu
sentence-transformers
openai
pydantic